+++ CHANGES.rst
@@ -11,7 +11,8 @@
   :meth:`pyudev.Enumerator.match_parent`
 - :attr:`pyudev.Device.children` requires udev version 172 now
 - #31: Added :meth:`pyudev.Enumerator.match_attribute`
-- Added ``nomatch`` argument to :meth:`pyudev.Enumerator.match_subsystem`
+- Added ``nomatch`` argument to :meth:`pyudev.Enumerator.match_subsystem` and
+  :meth:`pyudev.Enumerator.match_attribute`
 
 
 0.12 (Aug 31, 2011)